projecktFLUX/requirements.txt
python-telegram-bot==20.7
flask==2.3.3
gunicorn==21.2.0
psycopg2-binary==2.9.7